Add DockerHelper to wave-utils for Nextflow integration
- Create DockerHelper in wave-utils with file-based utility methods:
  - condaFileFromPath, condaFileFromPackages, condaPackagesToList
  - condaPackagesToCondaYaml, condaEnvironmentToCondaYaml
- Rename DockerHelper to TemplateUtils in main src (template-based methods)
- Remove duplicated methods from TemplateUtils (now in DockerHelper)
- Update ContainerHelper to import from DockerHelper

This separates concerns:
- DockerHelper (wave-utils): File-based utilities for Nextflow CLI
- TemplateUtils (wave main): Template-based Dockerfile generation
