+++ src/main/java/io/seqera/wave/util/TemplateUtils.java
@@ -20,10 +20,7 @@ package io.seqera.wave.util;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URL;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -31,101 +28,13 @@ import java.util.stream.Collectors;
 import io.seqera.wave.config.CondaOpts;
 import io.seqera.wave.config.PixiOpts;
 import org.apache.commons.lang3.StringUtils;
-import org.yaml.snakeyaml.DumperOptions;
-import org.yaml.snakeyaml.Yaml;
-import org.yaml.snakeyaml.representer.Representer;
 
 /**
  * Helper class to create Dockerfile for Conda package manager
  *
  * @author Paolo Di Tommaso <paolo.ditommaso@gmail.com>
  */
-public class DockerHelper {
-
-    static public List<String> condaPackagesToList(String packages) {
-        if (packages == null || packages.isEmpty())
-            return null;
-        return Arrays
-                .stream(packages.split(" "))
-                .filter(it -> !StringUtils.isEmpty(it))
-                .map(it -> trim0(it)).collect(Collectors.toList());
-    }
-
-    protected static String trim0(String value) {
-        if( value==null )
-            return null;
-        value = value.trim();
-        while( value.startsWith("'") && value.endsWith("'") )
-            value = value.substring(1,value.length()-1);
-        while( value.startsWith("\"") && value.endsWith("\"") )
-            value = value.substring(1,value.length()-1);
-        return value;
-    }
-
-    public static String condaPackagesToCondaYaml(String packages, List<String> channels) {
-        if( packages==null || packages.isBlank() )
-            return null;
-
-        final List<Object> deps = new ArrayList<>(20);
-        final List<Object> condaPackages = new ArrayList<>(10);
-        final List<Object> pipPackages = new ArrayList<>(10);
-        // split conda package by pip prefixed packages
-        for( String it : condaPackagesToList(packages) ) {
-            if( it.startsWith("pip::") )
-                throw new IllegalArgumentException(String.format("Invalid pip prefix - Likely you want to use '%s' instead of '%s'", it.replaceAll(":+",":"), it));
-            else if( it.startsWith("pip:") )
-                pipPackages.add(it.substring(4));
-            else
-                condaPackages.add(it);
-        }
-
-        // add all conda packages
-        if( !condaPackages.isEmpty() )
-            deps.addAll(condaPackages);
-
-        // add all pip packages
-        if( !pipPackages.isEmpty() ) {
-            deps.add("pip");
-            deps.add(Map.of("pip", pipPackages));
-        }
-
-        // add the channels
-        final Map<String, Object> conda = new LinkedHashMap<>();
-        if (channels != null && !channels.isEmpty() ) {
-            conda.put("channels", channels);
-        }
-
-        // assemble the final yaml
-        conda.put("dependencies", deps);
-        return dumpCondaYaml(conda);
-    }
-
-    static private String dumpCondaYaml(Map<String, Object> conda) {
-        DumperOptions dumperOpts = new DumperOptions();
-        dumperOpts.setPrettyFlow(false); // Disable pretty formatting
-        dumperOpts.setDefaultFlowStyle(DumperOptions.FlowStyle.BLOCK); // Use block style
-        return new Yaml(new Representer(dumperOpts), dumperOpts).dump(conda);
-    }
-
-    public static String condaEnvironmentToCondaYaml(String env, List<String> channels) {
-        final Yaml yaml = new Yaml();
-        // 1. parse the file
-        Map<String,Object> root = yaml.load(env);
-        // 2. append channels
-        if( channels!=null ) {
-            List<String> channels0 = (List<String>)root.get("channels");
-            if( channels0==null ) {
-                channels0 = new ArrayList<>();
-                root.put("channels", channels0);
-            }
-            for( String it : channels ) {
-                if( !channels0.contains(it) )
-                    channels0.add(it);
-            }
-        }
-        // 3. return it
-        return dumpCondaYaml(root);
-    }
+public class TemplateUtils {
 
     static public String condaPackagesToDockerFile(String packages, List<String> condaChannels, CondaOpts opts) {
         return condaPackagesTemplate0(
@@ -260,7 +169,7 @@ public class DockerHelper {
     }
 
     static private String renderTemplate0(String templatePath, Map<String,String> binding, List<String> ignore) {
-        final URL template = DockerHelper.class.getResource(templatePath);
+        final URL template = TemplateUtils.class.getResource(templatePath);
         if( template==null )
             throw new IllegalStateException(String.format("Unable to load template '%s' from classpath", templatePath));
         try {
