+++ src/test/groovy/io/seqera/wave/util/TemplateUtilsTest.groovy
@@ -25,218 +25,14 @@ import spock.lang.Specification
  *
  * @author Paolo Di Tommaso <paolo.ditommaso@gmail.com>
  */
-class DockerHelperTest extends Specification {
-
-    def 'should trim a string' () {
-        expect:
-        DockerHelper.trim0(STR) == EXPECTED
-
-        where:
-        STR         | EXPECTED
-        null        | null
-        "foo"       | "foo"
-        " foo  "    | "foo"
-        "'foo"      | "'foo"
-        '"foo'      | '"foo'
-        and:
-        "'foo'"     | "foo"
-        "''foo''"   | "foo"
-        " 'foo' "   | "foo"
-        " ' foo ' " | " foo "
-        and:
-        '"foo"'     | 'foo'
-        '""foo""'   | 'foo'
-        ' "foo" '   | 'foo'
-    }
-
-    def 'should convert conda packages to list' () {
-        expect:
-        DockerHelper.condaPackagesToList(STR) == EXPECTED
-
-        where:
-        STR                 | EXPECTED
-        "foo"               | ["foo"]
-        "foo bar"           | ["foo", "bar"]
-        "foo 'bar'"         | ["foo", "bar"]
-        "foo    'bar'  "    | ["foo", "bar"]
-    }
-
-    def 'should create conda yaml file' () {
-        expect:
-        DockerHelper.condaPackagesToCondaYaml("foo=1.0 'bar>=2.0'", null)
-            ==  '''\
-                dependencies:
-                - foo=1.0
-                - bar>=2.0
-                '''.stripIndent(true)
-
-
-        DockerHelper.condaPackagesToCondaYaml('foo=1.0 bar=2.0', ['channel_a','channel_b'] )
-                ==  '''\
-                channels:
-                - channel_a
-                - channel_b
-                dependencies:
-                - foo=1.0
-                - bar=2.0
-                '''.stripIndent(true)
-
-        DockerHelper.condaPackagesToCondaYaml('foo=1.0 bar=2.0 pip:numpy pip:pandas', ['channel_a','channel_b'] )
-                ==  '''\
-                channels:
-                - channel_a
-                - channel_b
-                dependencies:
-                - foo=1.0
-                - bar=2.0
-                - pip
-                - pip:
-                  - numpy
-                  - pandas
-                '''.stripIndent(true)
-    }
-
-    def 'should return an error when using invalid pip prefix' () {
-        when:
-        DockerHelper.condaPackagesToCondaYaml('pip::numpy', [] )
-        then:
-        def e = thrown(IllegalArgumentException)
-        e.message == "Invalid pip prefix - Likely you want to use 'pip:numpy' instead of 'pip::numpy'"
-    }
-
-    def 'should map pip packages to conda yaml' () {
-        expect:
-        DockerHelper.condaPackagesToCondaYaml('pip:numpy pip:panda pip:matplotlib', ['defaults']) ==
-                '''\
-                channels:
-                - defaults
-                dependencies:
-                - pip
-                - pip:
-                  - numpy
-                  - panda
-                  - matplotlib
-                '''.stripIndent()
-
-        and:
-        DockerHelper.condaPackagesToCondaYaml(null, ['foo']) == null
-        DockerHelper.condaPackagesToCondaYaml('  ', ['foo']) == null
-    }
-
-    def 'should add conda packages to conda yaml /1' () {
-        given:
-        def text = '''\
-         dependencies:
-         - foo=1.0
-         - bar=2.0
-        '''.stripIndent(true)
-
-        when:
-        def result = DockerHelper.condaEnvironmentToCondaYaml(text, null)
-        then:
-        result == '''\
-         dependencies:
-         - foo=1.0
-         - bar=2.0
-        '''.stripIndent(true)
-
-        when:
-        result = DockerHelper.condaEnvironmentToCondaYaml(text, ['ch1', 'ch2'])
-        then:
-        result == '''\
-             dependencies:
-             - foo=1.0
-             - bar=2.0
-             channels:
-             - ch1
-             - ch2
-            '''.stripIndent(true)
-    }
-
-    def 'should add conda packages to conda yaml /2' () {
-        given:
-        def text = '''\
-         dependencies:
-         - foo=1.0
-         - bar=2.0
-         channels:
-         - hola
-         - ciao
-        '''.stripIndent(true)
-
-        when:
-        def result = DockerHelper.condaEnvironmentToCondaYaml(text, null)
-        then:
-        result == '''\
-         dependencies:
-         - foo=1.0
-         - bar=2.0
-         channels:
-         - hola
-         - ciao
-        '''.stripIndent(true)
-
-        when:
-        result = DockerHelper.condaEnvironmentToCondaYaml(text, ['ch1', 'ch2'])
-        then:
-        result == '''\
-             dependencies:
-             - foo=1.0
-             - bar=2.0
-             channels:
-             - hola
-             - ciao
-             - ch1
-             - ch2
-            '''.stripIndent(true)
-    }
-
-    def 'should add conda packages to conda yaml /3' () {
-        given:
-        def text = '''\
-         channels:
-         - hola
-         - ciao
-        '''.stripIndent(true)
-
-        when:
-        def result = DockerHelper.condaEnvironmentToCondaYaml(text, null)
-        then:
-        result == '''\
-         channels:
-         - hola
-         - ciao
-        '''.stripIndent(true)
-
-        when:
-        result = DockerHelper.condaEnvironmentToCondaYaml(text, ['ch1', 'ch2'])
-        then:
-        result == '''\
-             channels:
-             - hola
-             - ciao
-             - ch1
-             - ch2
-            '''.stripIndent(true)
-
-        when:
-        result = DockerHelper.condaEnvironmentToCondaYaml(text, ['bioconda'])
-        then:
-        result == '''\
-             channels:
-             - hola
-             - ciao
-             - bioconda
-            '''.stripIndent(true)
-    }
-
+class TemplateUtilsTest extends Specification {
 
     def 'should create dockerfile content from conda file' () {
         given:
         def CONDA_OPTS = new CondaOpts([basePackages: 'foo::bar'])
 
         expect:
-        DockerHelper.condaFileToDockerFile(CONDA_OPTS)== '''\
+        TemplateUtils.condaFileToDockerFile(CONDA_OPTS)== '''\
                 FROM mambaorg/micromamba:1.5.10-noble
                 COPY --chown=$MAMBA_USER:$MAMBA_USER conda.yml /tmp/conda.yml
                 RUN micromamba install -y -n base -f /tmp/conda.yml \\
@@ -254,7 +50,7 @@ class DockerHelperTest extends Specification {
     def 'should create dockerfile content from conda file and base packages' () {
 
         expect:
-        DockerHelper.condaFileToDockerFile(new CondaOpts([:]))== '''\
+        TemplateUtils.condaFileToDockerFile(new CondaOpts([:]))== '''\
                 FROM mambaorg/micromamba:1.5.10-noble
                 COPY --chown=$MAMBA_USER:$MAMBA_USER conda.yml /tmp/conda.yml
                 RUN micromamba install -y -n base -f /tmp/conda.yml \\
@@ -275,7 +71,7 @@ class DockerHelperTest extends Specification {
         def PACKAGES = 'bwa=0.7.15 salmon=1.1.1'
         def CHANNELS = ['conda-forge', 'defaults']
         expect:
-        DockerHelper.condaPackagesToDockerFile(PACKAGES, CHANNELS, new CondaOpts([:])) == '''\
+        TemplateUtils.condaPackagesToDockerFile(PACKAGES, CHANNELS, new CondaOpts([:])) == '''\
                 FROM mambaorg/micromamba:1.5.10-noble
                 RUN \\
                     micromamba install -y -n base -c conda-forge -c defaults bwa=0.7.15 salmon=1.1.1 \\
@@ -297,7 +93,7 @@ class DockerHelperTest extends Specification {
         def PACKAGES = 'bwa=0.7.15 salmon=1.1.1'
 
         expect:
-        DockerHelper.condaPackagesToDockerFile(PACKAGES, CHANNELS, CONDA_OPTS) == '''\
+        TemplateUtils.condaPackagesToDockerFile(PACKAGES, CHANNELS, CONDA_OPTS) == '''\
                 FROM mambaorg/micromamba:1.5.10-noble
                 RUN \\
                     micromamba install -y -n base -c conda-forge -c defaults bwa=0.7.15 salmon=1.1.1 \\
@@ -318,7 +114,7 @@ class DockerHelperTest extends Specification {
         def PACKAGES = 'bwa=0.7.15 salmon=1.1.1'
 
         expect:
-        DockerHelper.condaPackagesToDockerFile(PACKAGES, CHANNELS, new CondaOpts([:])) == '''\
+        TemplateUtils.condaPackagesToDockerFile(PACKAGES, CHANNELS, new CondaOpts([:])) == '''\
                 FROM mambaorg/micromamba:1.5.10-noble
                 RUN \\
                     micromamba install -y -n base -c foo -c bar bwa=0.7.15 salmon=1.1.1 \\
@@ -340,7 +136,7 @@ class DockerHelperTest extends Specification {
         def PACKAGES = 'bwa=0.7.15 salmon=1.1.1'
 
         expect:
-        DockerHelper.condaPackagesToDockerFile(PACKAGES, CHANNELS, new CondaOpts(CONDA_OPTS)) == '''\
+        TemplateUtils.condaPackagesToDockerFile(PACKAGES, CHANNELS, new CondaOpts(CONDA_OPTS)) == '''\
                 FROM my-base:123
                 RUN \\
                     micromamba install -y -n base -c conda-forge -c defaults bwa=0.7.15 salmon=1.1.1 \\
@@ -365,7 +161,7 @@ class DockerHelperTest extends Specification {
         def PACKAGES = 'https://foo.com/some/conda-lock.yml'
 
         expect:
-        DockerHelper.condaPackagesToDockerFile(PACKAGES, CHANNELS, new CondaOpts(OPTS)) == '''\
+        TemplateUtils.condaPackagesToDockerFile(PACKAGES, CHANNELS, new CondaOpts(OPTS)) == '''\
                 FROM my-base:123
                 RUN \\
                     micromamba install -y -n base -c conda-forge -c defaults -f https://foo.com/some/conda-lock.yml \\
@@ -392,7 +188,7 @@ class DockerHelperTest extends Specification {
         def CONDA_OPTS = new CondaOpts([basePackages: 'foo::bar=1.0'])
 
         expect:
-        DockerHelper.condaFileToSingularityFile(CONDA_OPTS)== '''\
+        TemplateUtils.condaFileToSingularityFile(CONDA_OPTS)== '''\
                 BootStrap: docker
                 From: mambaorg/micromamba:1.5.10-noble
                 %files
@@ -413,7 +209,7 @@ class DockerHelperTest extends Specification {
     def 'should create singularity content from conda file and base packages' () {
 
         expect:
-        DockerHelper.condaFileToSingularityFile(new CondaOpts([:]))== '''\
+        TemplateUtils.condaFileToSingularityFile(new CondaOpts([:]))== '''\
                 BootStrap: docker
                 From: mambaorg/micromamba:1.5.10-noble
                 %files
@@ -437,7 +233,7 @@ class DockerHelperTest extends Specification {
         def PACKAGES = 'bwa=0.7.15 salmon=1.1.1'
         def CHANNELS = ['conda-forge', 'defaults']
         expect:
-        DockerHelper.condaPackagesToSingularityFile(PACKAGES, CHANNELS, new CondaOpts([:])) == '''\
+        TemplateUtils.condaPackagesToSingularityFile(PACKAGES, CHANNELS, new CondaOpts([:])) == '''\
                 BootStrap: docker
                 From: mambaorg/micromamba:1.5.10-noble
                 %post
@@ -460,7 +256,7 @@ class DockerHelperTest extends Specification {
         def PACKAGES = 'bwa=0.7.15 salmon=1.1.1'
 
         expect:
-        DockerHelper.condaPackagesToSingularityFile(PACKAGES, CHANNELS, CONDA_OPTS) == '''\
+        TemplateUtils.condaPackagesToSingularityFile(PACKAGES, CHANNELS, CONDA_OPTS) == '''\
                 BootStrap: docker
                 From: mambaorg/micromamba:1.5.10-noble
                 %post
@@ -482,7 +278,7 @@ class DockerHelperTest extends Specification {
         def PACKAGES = 'bwa=0.7.15 salmon=1.1.1'
 
         expect:
-        DockerHelper.condaPackagesToSingularityFile(PACKAGES, CHANNELS, new CondaOpts([:])) == '''\
+        TemplateUtils.condaPackagesToSingularityFile(PACKAGES, CHANNELS, new CondaOpts([:])) == '''\
                 BootStrap: docker
                 From: mambaorg/micromamba:1.5.10-noble
                 %post
@@ -505,7 +301,7 @@ class DockerHelperTest extends Specification {
         def PACKAGES = 'bwa=0.7.15 salmon=1.1.1'
 
         expect:
-        DockerHelper.condaPackagesToSingularityFile(PACKAGES, CHANNELS, new CondaOpts(CONDA_OPTS)) == '''\
+        TemplateUtils.condaPackagesToSingularityFile(PACKAGES, CHANNELS, new CondaOpts(CONDA_OPTS)) == '''\
                 BootStrap: docker
                 From: my-base:123
                 %post
@@ -532,7 +328,7 @@ class DockerHelperTest extends Specification {
         def PACKAGES = 'https://foo.com/some/conda-lock.yml'
 
         expect:
-        DockerHelper.condaPackagesToSingularityFile(PACKAGES, CHANNELS, new CondaOpts(OPTS)) == '''\
+        TemplateUtils.condaPackagesToSingularityFile(PACKAGES, CHANNELS, new CondaOpts(OPTS)) == '''\
                 BootStrap: docker
                 From: my-base:123
                 %post
@@ -555,7 +351,7 @@ class DockerHelperTest extends Specification {
         def PIXI_OPTS = new PixiOpts([basePackages: 'foo::bar'])
 
         expect:
-        DockerHelper.condaFileToDockerFileUsingPixi(PIXI_OPTS)== '''\
+        TemplateUtils.condaFileToDockerFileUsingPixi(PIXI_OPTS)== '''\
                 FROM ghcr.io/prefix-dev/pixi:0.59.0-noble AS build
 
                 COPY conda.yml /opt/wave/conda.yml
@@ -586,7 +382,7 @@ class DockerHelperTest extends Specification {
 
     def 'should create dockerfile content from conda file using pixi with default options' () {
         expect:
-        DockerHelper.condaFileToDockerFileUsingPixi(new PixiOpts([:])) == '''\
+        TemplateUtils.condaFileToDockerFileUsingPixi(new PixiOpts([:])) == '''\
                 FROM ghcr.io/prefix-dev/pixi:0.59.0-noble AS build
 
                 COPY conda.yml /opt/wave/conda.yml
@@ -620,7 +416,7 @@ class DockerHelperTest extends Specification {
         def PIXI_OPTS = new PixiOpts([baseImage: 'debian:12'])
 
         expect:
-        DockerHelper.condaFileToDockerFileUsingPixi(PIXI_OPTS)== '''\
+        TemplateUtils.condaFileToDockerFileUsingPixi(PIXI_OPTS)== '''\
                 FROM ghcr.io/prefix-dev/pixi:0.59.0-noble AS build
 
                 COPY conda.yml /opt/wave/conda.yml
@@ -649,84 +445,6 @@ class DockerHelperTest extends Specification {
                 '''.stripIndent()
     }
 
-    def 'should handle empty packages string' () {
-        expect:
-        DockerHelper.condaPackagesToCondaYaml('', ['conda-forge']) == null
-        DockerHelper.condaPackagesToCondaYaml(null, ['conda-forge']) == null
-    }
-
-    def 'should handle only pip packages' () {
-        expect:
-        DockerHelper.condaPackagesToCondaYaml('pip:requests pip:flask', null) ==
-                '''\
-                dependencies:
-                - pip
-                - pip:
-                  - requests
-                  - flask
-                '''.stripIndent()
-    }
-
-    def 'should handle packages with complex version specifications' () {
-        expect:
-        DockerHelper.condaPackagesToCondaYaml('numpy>=1.21.0,<2.0.0 pandas==1.5.3', ['conda-forge']) ==
-                '''\
-                channels:
-                - conda-forge
-                dependencies:
-                - numpy>=1.21.0,<2.0.0
-                - pandas==1.5.3
-                '''.stripIndent()
-    }
-
-    def 'should handle multiple pip packages correctly' () {
-        expect:
-        DockerHelper.condaPackagesToCondaYaml('pip:numpy==1.21.0 pip:pandas>=1.3.0 pip:matplotlib', ['defaults']) ==
-                '''\
-                channels:
-                - defaults
-                dependencies:
-                - pip
-                - pip:
-                  - numpy==1.21.0
-                  - pandas>=1.3.0
-                  - matplotlib
-                '''.stripIndent()
-    }
-
-    def 'should handle mixed conda and pip packages' () {
-        expect:
-        DockerHelper.condaPackagesToCondaYaml('bwa=0.7.15 pip:numpy pip:pandas salmon=1.1.1', ['conda-forge']) ==
-                '''\
-                channels:
-                - conda-forge
-                dependencies:
-                - bwa=0.7.15
-                - salmon=1.1.1
-                - pip
-                - pip:
-                  - numpy
-                  - pandas
-                '''.stripIndent()
-    }
-
-    def 'should handle packages with version specifiers' () {
-        given:
-        def packages = 'bwa>=0.7.15 salmon<=1.1.1 samtools==1.10'
-        def channels = ['bioconda']
-
-        expect:
-        DockerHelper.condaPackagesToCondaYaml(packages, channels) ==
-                '''\
-                channels:
-                - bioconda
-                dependencies:
-                - bwa>=0.7.15
-                - salmon<=1.1.1
-                - samtools==1.10
-                '''.stripIndent()
-    }
-
     /* *********************************************************************************
      * Micromamba v2 template tests (multi-stage builds)
      * *********************************************************************************/
@@ -740,7 +458,7 @@ class DockerHelperTest extends Specification {
         ])
 
         expect:
-        DockerHelper.condaFileToDockerFileUsingV2(CONDA_OPTS) == '''\
+        TemplateUtils.condaFileToDockerFileUsingV2(CONDA_OPTS) == '''\
                 FROM mambaorg/micromamba:2.1.1 AS build
                 COPY --chown=$MAMBA_USER:$MAMBA_USER conda.yml /tmp/conda.yml
                 RUN micromamba install -y -n base -f /tmp/conda.yml \\
@@ -761,7 +479,7 @@ class DockerHelperTest extends Specification {
 
     def 'should create dockerfile using micromamba v2 template with default options' () {
         expect:
-        DockerHelper.condaFileToDockerFileUsingV2(new CondaOpts([:])) == '''\
+        TemplateUtils.condaFileToDockerFileUsingV2(new CondaOpts([:])) == '''\
                 FROM mambaorg/micromamba:1.5.10-noble AS build
                 COPY --chown=$MAMBA_USER:$MAMBA_USER conda.yml /tmp/conda.yml
                 RUN micromamba install -y -n base -f /tmp/conda.yml \\
@@ -791,7 +509,7 @@ class DockerHelperTest extends Specification {
         ])
 
         expect:
-        DockerHelper.condaPackagesToDockerFileUsingV2(PACKAGES, CHANNELS, CONDA_OPTS) == '''\
+        TemplateUtils.condaPackagesToDockerFileUsingV2(PACKAGES, CHANNELS, CONDA_OPTS) == '''\
                 FROM mambaorg/micromamba:2.1.1 AS build
                 RUN \\
                     micromamba install -y -n base -c conda-forge -c bioconda bwa=0.7.15 salmon=1.1.1 \\
@@ -821,7 +539,7 @@ class DockerHelperTest extends Specification {
         ])
 
         expect:
-        DockerHelper.condaPackagesToDockerFileUsingV2(PACKAGES, CHANNELS, CONDA_OPTS) == '''\
+        TemplateUtils.condaPackagesToDockerFileUsingV2(PACKAGES, CHANNELS, CONDA_OPTS) == '''\
                 FROM mambaorg/micromamba:2.1.1 AS build
                 RUN \\
                     micromamba install -y -n base -c conda-forge numpy pandas \\
@@ -851,7 +569,7 @@ class DockerHelperTest extends Specification {
         ])
 
         when:
-        def result = DockerHelper.condaPackagesToDockerFileUsingV2(PACKAGES, CHANNELS, CONDA_OPTS)
+        def result = TemplateUtils.condaPackagesToDockerFileUsingV2(PACKAGES, CHANNELS, CONDA_OPTS)
 
         then:
         result.contains('FROM mambaorg/micromamba:2.1.1 AS build')
@@ -870,7 +588,7 @@ class DockerHelperTest extends Specification {
         ])
 
         when:
-        def result = DockerHelper.condaPackagesToDockerFileUsingV2(PACKAGES, CHANNELS, CONDA_OPTS)
+        def result = TemplateUtils.condaPackagesToDockerFileUsingV2(PACKAGES, CHANNELS, CONDA_OPTS)
 
         then:
         result.contains('-f https://foo.com/some/conda-lock.yml')
@@ -891,7 +609,7 @@ class DockerHelperTest extends Specification {
         ])
 
         expect:
-        DockerHelper.condaFileToSingularityFileUsingPixi(PIXI_OPTS) == '''\
+        TemplateUtils.condaFileToSingularityFileUsingPixi(PIXI_OPTS) == '''\
                 BootStrap: docker
                 From: ghcr.io/prefix-dev/pixi:latest
                 Stage: build
@@ -927,7 +645,7 @@ class DockerHelperTest extends Specification {
 
     def 'should create singularityfile using pixi v1 template with default options' () {
         expect:
-        DockerHelper.condaFileToSingularityFileUsingPixi(new PixiOpts([:])) == '''\
+        TemplateUtils.condaFileToSingularityFileUsingPixi(new PixiOpts([:])) == '''\
                 BootStrap: docker
                 From: ghcr.io/prefix-dev/pixi:0.59.0-noble
                 Stage: build
@@ -970,7 +688,7 @@ class DockerHelperTest extends Specification {
         ])
 
         expect:
-        DockerHelper.condaFileToSingularityFileUsingPixi(PIXI_OPTS) == '''\
+        TemplateUtils.condaFileToSingularityFileUsingPixi(PIXI_OPTS) == '''\
                 BootStrap: docker
                 From: ghcr.io/prefix-dev/pixi:0.35.0
                 Stage: build
@@ -1011,7 +729,7 @@ class DockerHelperTest extends Specification {
         ])
 
         when:
-        def result = DockerHelper.condaFileToDockerFileUsingPixi(PIXI_OPTS)
+        def result = TemplateUtils.condaFileToDockerFileUsingPixi(PIXI_OPTS)
 
         then:
         result.contains('pixi add conda-forge::procps-ng')
@@ -1027,7 +745,7 @@ class DockerHelperTest extends Specification {
         ])
 
         when:
-        def result = DockerHelper.condaFileToSingularityFileUsingPixi(PIXI_OPTS)
+        def result = TemplateUtils.condaFileToSingularityFileUsingPixi(PIXI_OPTS)
 
         then:
         result.contains('pixi add conda-forge::bash')
@@ -1045,7 +763,7 @@ class DockerHelperTest extends Specification {
         ])
 
         expect:
-        DockerHelper.condaFileToSingularityFileV2(CONDA_OPTS) == '''\
+        TemplateUtils.condaFileToSingularityFileV2(CONDA_OPTS) == '''\
                 BootStrap: docker
                 From: mambaorg/micromamba:2.1.1
                 Stage: build
@@ -1078,7 +796,7 @@ class DockerHelperTest extends Specification {
 
     def 'should create singularityfile using micromamba v2 template with default options' () {
         expect:
-        DockerHelper.condaFileToSingularityFileV2(new CondaOpts([:])) == '''\
+        TemplateUtils.condaFileToSingularityFileV2(new CondaOpts([:])) == '''\
                 BootStrap: docker
                 From: mambaorg/micromamba:1.5.10-noble
                 Stage: build
@@ -1119,7 +837,7 @@ class DockerHelperTest extends Specification {
         ])
 
         expect:
-        DockerHelper.condaPackagesToSingularityFileV2(PACKAGES, CHANNELS, CONDA_OPTS) == '''\
+        TemplateUtils.condaPackagesToSingularityFileV2(PACKAGES, CHANNELS, CONDA_OPTS) == '''\
                 BootStrap: docker
                 From: mambaorg/micromamba:2.1.1
                 Stage: build
@@ -1153,7 +871,7 @@ class DockerHelperTest extends Specification {
         ])
 
         expect:
-        DockerHelper.condaPackagesToSingularityFileV2(PACKAGES, CHANNELS, CONDA_OPTS) == '''\
+        TemplateUtils.condaPackagesToSingularityFileV2(PACKAGES, CHANNELS, CONDA_OPTS) == '''\
                 BootStrap: docker
                 From: mambaorg/micromamba:2.1.1
                 Stage: build
@@ -1187,7 +905,7 @@ class DockerHelperTest extends Specification {
         ])
 
         when:
-        def result = DockerHelper.condaPackagesToSingularityFileV2(PACKAGES, CHANNELS, CONDA_OPTS)
+        def result = TemplateUtils.condaPackagesToSingularityFileV2(PACKAGES, CHANNELS, CONDA_OPTS)
 
         then:
         result.contains('From: mambaorg/micromamba:2.1.1')
@@ -1207,7 +925,7 @@ class DockerHelperTest extends Specification {
         ])
 
         when:
-        def result = DockerHelper.condaPackagesToSingularityFileV2(PACKAGES, CHANNELS, CONDA_OPTS)
+        def result = TemplateUtils.condaPackagesToSingularityFileV2(PACKAGES, CHANNELS, CONDA_OPTS)
 
         then:
         result.contains('-f https://foo.com/some/conda-lock.yml')
